CWL: accept one input port to accept from multiple output ports (reasoner may not be right, but works)
CWLProv sometimes contain one input port reading from multiple outputs (reading one data with prov:hadMember where the members are from multiple outputs).  This may be due to automatic parallelisation.
The queries have been updated to accept such scenarios. It should be correct.
But the reasoner was built with the assumption that "one input reads from one output". This makes it a potential problem. But the tests shows it works and propagates rules from all inputs. This may be related to the "merge" behaviour in the Python code when reading inputs.